requirements.txt
beautifulsoup4==4.9.3
bs4==0.0.1
jupyter==1.0.0
lxml==4.6.1
pandas==1.1.4
requests==2.24.0
psycopg2-binary==2.8.6
SQLAlchemy==1.3.20
